+++ packages/camera/camera/android/src/test/java/io/flutter/plugins/camera/CameraTest.java
@@ -22,11 +22,15 @@ import android.hardware.camera2.CameraAccessException;
 import android.hardware.camera2.CameraCaptureSession;
 import android.hardware.camera2.CameraMetadata;
 import android.hardware.camera2.CaptureRequest;
+import android.hardware.camera2.params.SessionConfiguration;
+import android.media.ImageReader;
 import android.media.MediaRecorder;
 import android.os.Build;
 import android.os.Handler;
 import android.os.HandlerThread;
+import android.view.Surface;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.lifecycle.LifecycleObserver;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.plugin.common.MethodChannel;
@@ -50,10 +54,38 @@ import io.flutter.plugins.camera.features.sensororientation.SensorOrientationFea
 import io.flutter.plugins.camera.features.zoomlevel.ZoomLevelFeature;
 import io.flutter.plugins.camera.utils.TestUtils;
 import io.flutter.view.TextureRegistry;
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.MockedStatic;
+
+class FakeCameraDeviceWrapper implements CameraDeviceWrapper {
+  final List<CaptureRequest.Builder> captureRequests;
+
+  FakeCameraDeviceWrapper(List<CaptureRequest.Builder> captureRequests) {
+    this.captureRequests = captureRequests;
+  }
+
+  @NonNull
+  @Override
+  public CaptureRequest.Builder createCaptureRequest(int var1) {
+    return captureRequests.remove(0);
+  }
+
+  @Override
+  public void createCaptureSession(SessionConfiguration config) {}
+
+  @Override
+  public void createCaptureSession(
+      @NonNull List<Surface> outputs,
+      @NonNull CameraCaptureSession.StateCallback callback,
+      @Nullable Handler handler) {}
+
+  @Override
+  public void close() {}
+}
 
 public class CameraTest {
   private CameraProperties mockCameraProperties;
@@ -799,6 +831,29 @@ public class CameraTest {
     camera.startBackgroundThread();
 
     verify(mockHandlerThread, times(1)).start();
+  }
+
+  @Test
+  public void onConverge_shouldTakePictureWithoutAbortingSession() throws CameraAccessException {
+    ArrayList<CaptureRequest.Builder> mockRequestBuilders = new ArrayList<>();
+    mockRequestBuilders.add(mock(CaptureRequest.Builder.class));
+    CameraDeviceWrapper fakeCamera = new FakeCameraDeviceWrapper(mockRequestBuilders);
+    // Stub out other features used by the flow.
+    TestUtils.setPrivateField(camera, "cameraDevice", fakeCamera);
+    TestUtils.setPrivateField(camera, "pictureImageReader", mock(ImageReader.class));
+    SensorOrientationFeature mockSensorOrientationFeature =
+        mockCameraFeatureFactory.createSensorOrientationFeature(mockCameraProperties, null, null);
+    DeviceOrientationManager mockDeviceOrientationManager = mock(DeviceOrientationManager.class);
+    when(mockSensorOrientationFeature.getDeviceOrientationManager())
+        .thenReturn(mockDeviceOrientationManager);
+
+    // Simulate a post-precapture flow.
+    camera.onConverged();
+    // A picture should be taken.
+    verify(mockCaptureSession, times(1)).capture(any(), any(), any());
+    // The session shuold not be aborted as part of this flow, as this breaks capture on some
+    // devices, and causes delays on others.
+    verify(mockCaptureSession, never()).abortCaptures();
   }
 
   private static class TestCameraFeatureFactory implements CameraFeatureFactory {