+++ packages/camera/camera/CHANGELOG.md
@@ -1,6 +1,16 @@
-## NEXT
+## 0.9.4+17
 
-* Minor iOS internal code cleanup related to queue helper functions.  
+* Removes obsolete information from README, and adds OS support table.
+
+## 0.9.4+16
+
+* Fixes a bug resulting in a `CameraAccessException` that prevents image
+  capture on some Android devices.
+
+## 0.9.4+15
+
+* Uses dispatch queue for pixel buffer synchronization on iOS.
+* Minor iOS internal code cleanup related to queue helper functions.
 
 ## 0.9.4+14